# openapi.yaml
openapi: 3.0.3
info:
  title: Smart Chat RAG Chatbot API
  description: |
    RESTful API for Smart Chat - an AI-powered RAG (Retrieval-Augmented Generation) chatbot platform
    with WordPress integration, multi-bot management, and human handoff capabilities.

    ## Features
    - Multi-tenant bot management
    - RAG-based responses using vector embeddings
    - WordPress plugin integration
    - Human agent handoff
    - Stripe billing integration
    - Rate limiting & abuse detection
    - Streaming responses (SSE)

    ## Authentication
    Most endpoints require authentication via NextAuth session cookies or API tokens.
    WordPress endpoints use custom token format: `userId:botId:secret`

  version: 1.0.0
  contact:
    name: API Support
    email: support@smartchat.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.smartchat.com/api
    description: Production server
...
security:
  - sessionAuth: []
  - tokenAuth: []
paths:
...
  /wordpress/validate-token:
    post:
      tags:
        - WordPress
      summary: Validate WordPress integration token
      description: Verify if a WordPress token is valid and active
      security:
        - tokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  example: "123e4567:abc-123:secret"
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    example: true
                  botId:
                    type: string
                    format: uuid
                  userId:
                    type: string
                    format: uuid
        '401':
          description: Token is invalid or inactive
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==========================================
  # AUTHENTICATION ENDPOINTS
  # ==========================================
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: WordPress integration token (userId:botId:secret)
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
        message:
          type: string
          example: "Detailed error description"
...